Documentation and replanning of task synchronization
Continuing to document CAN component, which led to a realization that my current control flow of data between tasks does not fully utilize the asynchronous nature of FreeRTOS tasks. Instead of funneling all interfaced component data through the puzzle component to send to the CAN component, those components should send their data to the CAN component directly. This will greatly reduce the complexity of the code, eliminate unnecessary bottlenecks in the puzzle component, and reduce the amount of instructions to execute overall.