CAN component logic flow
Data flow between CAN component and the rest of the components has been programmed for incoming requests. Need to determine a flow for components to send outgoing requests without an external command. Double check rx_task is correct. Make sure the code logic flow on the Google Doc is up to date.